requirements-cuda.txt
# Patchwork AlphaZero Dependencies (GPU/CUDA Version)
# For systems with NVIDIA GPU and CUDA support

# Core numerical libraries
numpy>=1.24.0,<2.0.0

# PyTorch with CUDA 12.1 support (compatible with CUDA 12.x and 13.x)
--index-url https://download.pytorch.org/whl/cu121
torch>=2.0.0
torchvision>=0.15.0

# Configuration and logging
pyyaml>=6.0
tensorboard>=2.12.0

# Data storage and serialization
h5py>=3.8.0

# Performance acceleration (optional but recommended)
numba>=0.57.0

# Testing and development
pytest>=7.3.0
pytest-cov>=4.1.0

# Code quality (optional)
black>=23.3.0
flake8>=6.0.0
mypy>=1.3.0
